Add not-simple AI action list (restores AHeal).
Simple list is a little too simple for a complete experience while
playing Reassembly, but is *much* better for a first look at AI modding.
Keeping both in place with an ifdef toggle.  Defaulting to not-simple so
the mod's more fun to play out of the box.

Not-simple list is from the boilerplate AI mod example.  With the
addition of the two new actions the weapon-engine-assist mod add.